Fix memory leak in request_state_transfer
Valgrind detected memory leak in request_state_transfer method.
The leak is possible only in the path when the server aborts. In such
a case memory allocated in prepare_state_request is not freed. As we end
up with abort() this is not critical issue. All happy paths free
the memory correctly.
The solution is to add missing free to make Valgrind happy.