PXC-4779: Code refresh for - PXC 8.0.44
https://perconadev.atlassian.net/browse/PXC-4779

Fixed compilation on ubuntu:noble (GCC 13.3.0) with MAINTAINER_MODE=ON.

Inside realloc we cache passed pointer in 'orig' variable, then realloc
original pointer, then use 'orig' in comparison and to erase from
allocated list (see the implementation). The compiler complains about
'orig' to be reused after free, which is correct, but suppressing the
warning does not solve the problem ultimately. It remains in the unit
test. Even if we suppress the warning for the whole realloc method it
does not solve the warning.
Most probably it is because realloc is inlined, and the compiler makes
'orig' an alias to the passed pointer, and as it is inlined, it sees
that it can be used in the caller function.
That's why we suppres the warning in the caller (unit test).

There is a clear ultimate solution for the warning, but it involves
erasing the original pointer from the list, then reallocating and then
reinserting the pointer. As realloc is a busy function, such an approach
seems to be not optimal.

Also disabling the waring in CMakeFile for the whole GCache seems to be
not a good idea.